[refac] cell 에서 WakeUpTimeView Extension이 활용 불가능한 부분을 수정함. (#69)
* [fix] DailyDiaryVC placeholder 문제 해결, 그 외 textView 관련 문제도 해결
body:
1. placeholder 가 다른데에 focus 갔다가 다시 돌아오면 Placeholder 가 사라지지 않는문제가 있어서 해결함
2. 직접 device test 를 했을 때 일기, 짧은 독서 뷰에서 글자를 쓴 후 화면 누르면 키보드 내려가는 기능이 없어서 추가함, @xwoud 의 코드 일부분 사용함.
3. [layout] 모든 뷰의 headerLabel 의 색깔이 지정 안되어 있었음. 설정해주었음.

* [style] 기상 인증 뷰 layout 완료
body:
dquote> 1. 이름과 같은 데이터전달이 될 매개변수는 기존에 @xwoud 님이 사용했던 변수 이름 그대로 userNick 으로 설정했음.
dquote> 2. uploadedPicture 는 UIImageView이며, 일단 test_img 파일을 assets 에 추가해서 연결해두었음. 실제 서버 연결에서는 회의 때 언급했던 kingfisher 의 image cashing 을 통해 이 부분에 데이터 전달을 하면
될 것으로 보임.
3. 22번째의 숫자도 매개변수화 해야할 것으로 보임.

* [refac] 숫자 변수화, letterSpacing 조정
body:
1. ~번째 의미있는 아침에서 숫자를 변수화 해주었음.
2. bodyUpperLabel 의 lineSpacing 을 조절하였음.. Zeplin 에 있는 것과 다른 숫자 개념인 것 같음. 일단 외관상 가장 비슷한 정도인 6 으로 line spacing 을 주었는데 이 부분에 대해서 이 extension 을 만든 본인이 검토를 해보도록 하겠습니다.

* [refac] 리뷰 피드백 반영
1. convention 수정
2. uploadedImage 데이터 전달 방식에 따라 변수화

* [style] 마이 피드 뷰 레이아
body:
1. rough 한 레이아웃만 잡았음
2. collection view 추가, nib 사용해보았음

* [refac] FeedImage model 추가, 연동 완료
body:
1. FeedImage struct 만들어두었음
2. @xwoud 의 피드백에 따라 ~번째 아침에서 숫자를 feedImageList 의 count로 설정하였음.

* [fix] collection view의 scroll view inset 조정
dquote> body:
dquote> 1. iphone SE 외의 큰 device 에서 collection view 아래 부분 safe area 가 비어있게 스크롤이 되는 문제 해결
2. UIScrollViewContentInsetAdjustment .never 로 설정함
3. 간단한 주석 추가함

* [refac] collectino view 부분 코드 정리

* [refac] random 숫자0,1에 따라 profile 색깔 변하도록 적용함
body:
1. random number 적용'
2. subview 에서는 profile view 원으로 만들고, 색깔 적용
3. 아래 setLayout 에서는 textcolor 적용
+ subview 에서 textlabel 색깔을 한꺼번에 바꾸려고 하니까 적용이 안되었었음.. 왜그런지에 대해서 궁금함.
4. FeedCollectionViewCell.swift 주석처리 함

* [upload] 매일 오전 ~시 ~분 기상 UIView 를 위한 새로운 Extension 추가
body:
1. 이름의 새로운 extension file added.
2. 상세 설명 읽기 바람
3. 마이 피드 뷰에도 새로운 extension 적용 완료

* [refac] cell 에서 WakeUpTimeView Extension이 활용 불가능한 부분을 수정함.
body:
1. 매개변수를 dateData 하나만 받는 것으로 수정함
2. dateData 하나만 받고, 매일 오전 9시 기상 과 같은 string 을 반환하고, 반환되는 string 의 count 값으로 각자 적용하는 view controller 에서 view width 분기처리를 하기로 결정함.
3. UIViewController extension 이 toast 도 있기 때문에 이 둘을  라는 한 파일로 통합함.
